Clases 14/01/22

+++ main.py
@@ -1,4 +1,6 @@
 import json
+from datetime import datetime
+from hashlib import sha256
 
 from flask import Flask, request
 
@@ -7,28 +9,51 @@ from ResponseModel import ResponseModel
 
 app = Flask(__name__)
 
+PASS = "abcd1234"
 
+def checkTokenAuth(tokenSHA256Request, USER, route):
 
+    passSHA256 = sha256(PASS.encode('utf-8')).hexdigest()
+    minutes = datetime.now().minute
+
+    tokenString = USER + route + passSHA256 + str(minutes)
+    tokenSHA256 = sha256(tokenString.encode('utf-8')).hexdigest()
+
+    print(tokenSHA256Request)
+    print(tokenSHA256)
+
+    if tokenSHA256 == tokenSHA256Request:
+        print('acceso correcto')
+        return True
+    else:
+        print('acceso denegado')
+        return False
 
 
 @app.route('/students', methods=['POST','PUT','DELETE','GET'])
 def students():
     print(request.json)
     response = ResponseModel()
+    tokenSHA256Request = request.authorization['password']
+    user = request.authorization['username']
+    route = request.json['route']
 
-    try:
-        if request.method == 'POST':
-            response = addStudent(request.json['data'])
-        elif request.method == 'GET':
-            response = getStudent(request.json['data'])
-        elif request.method == 'PUT':
-            response = updateStudent(request.json['data'])
-        elif request.method == 'DELETE':
-            response = deleteStudent(request.json['data'])
+    if checkTokenAuth(tokenSHA256Request, user, route):
+        try:
+            if request.method == 'POST':
+                response = addStudent(request.json['data'])
+            elif request.method == 'GET':
+                response = getStudent(request.json['data'])
+            elif request.method == 'PUT':
+                response = updateStudent(request.json['data'])
+            elif request.method == 'DELETE':
+                response = deleteStudent(request.json['data'])
 
 
-    except Exception as e:
-        print(e)
+        except Exception as e:
+            print(e)
+    else:
+        response.data = 'NO TIENES ACCESO'
 
     return json.dumps(response.__dict__)
 
@@ -46,7 +71,7 @@ def getStudent(_idE):
     if _idE == 'all':
         response = DBHandler().obtenerEstudiantes()
     else:
-        response = DBHandler().obtenerEstudiante()
+        response = DBHandler().obtenerEstudiante(_idE)
 
     return response
